Mini-Project-Sem6-main/campus-connect-backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
flask-cors==4.0.0
python-dotenv==1.0.0